Program core
GUI is currently not working, but the backbone of the program is. It can
create a new project using addScript(name) and then test the project
using runScript(name). Note that files in scripts/*/input/ and
scripts/*/output/ must be formatted as provided on the hackerrank
website (input*.txt and output*.txt)